+++ apps/web/app/(use-page-wrapper)/settings/(settings-layout)/organizations/teams/other/(main-page)/loading.tsx
@@ -1,0 +1,5 @@
+import { SkeletonLoader } from "./skeleton";
+
+export default function Loading() {
+  return <SkeletonLoader />;
+}
