+++ apps/web/app/(use-page-wrapper)/settings/(settings-layout)/organizations/teams/other/(main-page)/skeleton.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import SkeletonLoaderTeamList from "@calcom/ee/teams/components/SkeletonloaderTeamList";
+import SettingsHeader from "@calcom/features/settings/appDir/SettingsHeader";
+import { useLocale } from "@calcom/lib/hooks/useLocale";
+
+export function SkeletonLoader() {
+  const { t } = useLocale();
+  return (
+    <SettingsHeader title={t("org_admin_other_teams")} description={t("org_admin_other_teams_description")}>
+      <SkeletonLoaderTeamList />
+    </SettingsHeader>
+  );
+}
