--- apps/web/app/(use-page-wrapper)/settings/(settings-layout)/organizations/teams/other/page.tsx
@@ -1,25 +1,0 @@
-import { _generateMetadata, getTranslate } from "app/_utils";
-
-import LegacyPage from "@calcom/features/ee/organizations/pages/settings/other-team-listing-view";
-import SettingsHeader from "@calcom/features/settings/appDir/SettingsHeader";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    (t) => t("org_admin_other_teams"),
-    (t) => t("org_admin_other_teams_description"),
-    undefined,
-    undefined,
-    "/settings/organizations/teams/other"
-  );
-
-const Page = async () => {
-  const t = await getTranslate();
-
-  return (
-    <SettingsHeader title={t("org_admin_other_teams")} description={t("org_admin_other_teams_description")}>
-      <LegacyPage />
-    </SettingsHeader>
-  );
-};
-
-export default Page;
