+++ packages/features/ee/organizations/pages/components/OtherTeamList.tsx
@@ -1,14 +1,16 @@
+"use client";
+
 import { useState } from "react";
 
 import { trackFormbricksAction } from "@calcom/lib/formbricks-client";
-import type { RouterOutputs } from "@calcom/trpc/react";
+import type { OrganizationRepository } from "@calcom/lib/server/repository/organization";
 import { trpc } from "@calcom/trpc/react";
 import { showToast } from "@calcom/ui/components/toast";
 
 import OtherTeamListItem from "./OtherTeamListItem";
 
 interface Props {
-  teams: RouterOutputs["viewer"]["organizations"]["listOtherTeams"];
+  teams: Awaited<ReturnType<typeof OrganizationRepository.findTeamsInOrgIamNotPartOf>>;
   pending?: boolean;
 }
