--- packages/features/ee/organizations/pages/settings/other-team-listing-view.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import { useLocale } from "@calcom/lib/hooks/useLocale";
-
-import { OtherTeamsListing } from "./../components/OtherTeamsListing";
-
-const OtherTeamListingView = (): React.ReactElement => {
-  const { t } = useLocale();
-  return (
-    <>
-      <OtherTeamsListing />
-    </>
-  );
-};
-
-export default OtherTeamListingView;
